Pull traversal logic out into fs::util mod.
