+++ zagreus-server/src/endpoint/asset.rs
@@ -1,4 +1,3 @@
-use std::fs::DirEntry;
 use std::path::{Path, PathBuf};
 
 use axum::body::Bytes;
@@ -9,6 +8,7 @@ use axum::Json;
 use serde_json::json;
 
 use crate::error::ZagreusError;
+use crate::fs;
 
 pub const ASSETS_FOLDER_NAME: &str = "assets";
 
@@ -27,7 +27,7 @@ pub(crate) async fn get_asset_filenames(
     templates_data_folder.push(ASSETS_FOLDER_NAME);
 
     let template_assets_folder = templates_data_folder;
-    let traversal_result = traverse(&template_assets_folder);
+    let traversal_result = fs::util::traverse(&template_assets_folder);
 
     match traversal_result {
         Ok(entries) => (StatusCode::OK, Json(json!(entries))),
@@ -42,49 +42,6 @@ pub(crate) async fn get_asset_filenames(
                 Json(json!("Could not read template assets.")),
             )
         }
-    }
-}
-
-// TODO: Consider factoring out into a file helper module.
-
-#[derive(Serialize)]
-#[serde(untagged)]
-enum DirEntryNode {
-    File {
-        name: String,
-    },
-    Dir {
-        name: String,
-        children: Vec<DirEntryNode>,
-    },
-}
-
-fn traverse(path: &Path) -> Result<Vec<DirEntryNode>, ZagreusError> {
-    let files = std::fs::read_dir(path)?;
-    Ok(files
-        .filter_map(|entry| entry.ok())
-        .map(|entry| {
-            let child_path = entry.path();
-            let child_name = get_filename(entry)?;
-            match child_path.is_file() {
-                true => Ok(DirEntryNode::File { name: child_name }),
-                false => Ok(DirEntryNode::Dir {
-                    name: child_name,
-                    children: traverse(&child_path)?,
-                }),
-            }
-        })
-        .filter_map(|entry: Result<DirEntryNode, ZagreusError>| entry.ok())
-        .collect())
-}
-
-fn get_filename(entry: DirEntry) -> Result<String, ZagreusError> {
-    match entry.file_name().into_string() {
-        Ok(filename) => Ok(filename),
-        Err(_) => Err(ZagreusError::from(format!(
-            "Failed to process filename: {:?}",
-            entry.file_name()
-        ))),
     }
 }
